backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# HTTP & Utils
httpx>=0.26.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Crypto (token encryption)
cryptography>=42.0.0

# Celery + Redis
celery[redis]>=5.3.6
redis>=5.0.0

# SSE (Server-Sent Events)
sse-starlette>=1.6.0

# Social media scraping
yt-dlp>=2023.12.30
instaloader>=4.10.0

# LLM
google-generativeai>=0.3.0

# Monitoring
rich>=13.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
